+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @templatical/core
 
+## 0.29.1
+
+### Patch Changes
+
+- @templatical/types@0.29.1
+
 ## 0.29.0
 
 ### Minor Changes