+++ packages/media-library/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @templatical/media-library
 
+## 0.29.1
+
+### Patch Changes
+
+- @templatical/core@0.29.1
+  - @templatical/types@0.29.1
+
 ## 0.29.0
 
 ### Patch Changes