+++ packages/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @templatical/types
 
+## 0.29.1
+
 ## 0.29.0
 
 ### Minor Changes